Bump terraform-root-modules to 0.43.0
This is quite a jump but needed for some recent fixes to do with `users`
and `iam`.

One of the changes to how IAM user accounts are managed is to have all IAM users centrally managed in `conf/users`.

vadzim + aliaksandr are not currently working on CP projects so those
users are removed and the other users are in the correct admin groups
across all accounts thanks to the `admin_groups` local, where all admin
groups across enabled accounts are read from SSM.

 are added to these groups as they are included in `admin_groups